refactor(corebundle): split bundle.go into per-concern files (D2 part 2)
Pure file move, zero callsite churn. Apply PR-A66 actual范式 (per-concern
file split) to cmd/corebundle/bundle.go (566 lines, 26 top symbols → 5 files
+ doc.go anchor):

- bundle_options.go (138): runtimeBaseOptions + defaultRuntimeOptions +
  consumerMiddlewares + configEventConsumerMiddleware + buildInternalAuthChain
  + newBootstrapFromOptions ([]bootstrap.Option assembly).

- bundle_assembly.go (59): buildAssembly + durabilityModeForTopology +
  buildConsumerBase (SharedDeps → kernel object construction).

- bundle_keyprovider.go (116): buildKeyProvider + keyProviderToTransformer +
  noKeyProvider sentinel + isNoKeyProvider (KeyProvider construction +
  no-key sentinel implementation).

- bundle_configcore_storage.go (180): ConfigCoreModuleConfig +
  ConfigCoreModuleResult + buildConfigCoreOpts + verifyConfigCorePGSchema +
  buildConfigCorePGStorage (configcore-specific storage adapter wiring).

- bundle_devtools.go (110): devtoolsOption + parseProjectWithTimeout +
  defaultDevtoolsParseTimeout (devtools catalog endpoint, J1 PR-A37).

- doc.go (2): package docstring anchor.

bundle.go deleted; no symbol survives renamed or shadowed. All 50+ existing
test references to these symbols (key_provider_test / outbox_wiring_test /
bundle_test / vault_readiness_wiring_test / etc.) resolve unchanged via
same-package access.

Closes D2 (PR-V1-SHARED-DEPS-SPLIT) part 2 of 2.

ref: runtime/bootstrap/options_assembly.go / phases_assembly.go — per-concern
file split prior art (PR-A66 BOOTSTRAP-STRUCT-DECOMPOSE)
