+++ cmd/corebundle/doc.go
@@ -1,0 +1,2 @@
+// Package main is the corebundle composition root.
+package main
